Make Network.Socket work on Windows
Consolidate the network stuff in one file.
