+++ libs/contrib/Network/Socket/Raw.idr
@@ -8,9 +8,6 @@ module Network.Socket.Raw
 import public Network.Socket.Data
 
 %include C "idris_net.h"
-%include C "sys/types.h"
-%include C "sys/socket.h"
-%include C "netdb.h"
 
 %access public export
