+++ rts/Makefile
@@ -2,7 +2,7 @@ include ../config.mk
 
 OBJS = idris_rts.o idris_heap.o idris_gc.o idris_gmp.o idris_bitstring.o \
        idris_opts.o idris_stats.o idris_utf8.o idris_stdfgn.o mini-gmp.o \
-       idris_buffer.o getline.o
+       idris_buffer.o getline.o idris_net.o
 HDRS = idris_rts.h idris_heap.h idris_gc.h idris_gmp.h idris_bitstring.h \
        idris_opts.h idris_stats.h mini-gmp.h idris_stdfgn.h idris_net.h \
        idris_buffer.h idris_utf8.h getline.h
@@ -20,10 +20,9 @@ endif
 endif
 
 ifeq ($(OS), windows)
-	OBJS += windows/idris_net.o windows/win_utils.o
+	OBJS += windows/win_utils.o
 else
 	CFLAGS += -fPIC
-	OBJS += idris_net.o
 endif
 
 LIBTARGET = libidris_rts.a
